backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pypdf
python-barcode
python-multipart
python-dotenv
supabase
Pillow